+++ packages/data-export/src/csv/export-to-csv.ts
@@ -1,58 +1,22 @@
+import { ExportOutput } from '../types';
+
 /**
- * Creates CSV structure from parsed analysis results.
+ * Creates CSV structure from export output.
  *
  * @param data - parsed analysis results
- * @param columnDelimiter - character used as column delimeter
- * @param lineDelimiter - character used to recognize line delimeter
+ * @param columnDelimiter - character used as column delimiter
+ * @param lineDelimiter - character used to recognize line delimiter
  * @return transformed data
  *
  */
-export const exportToCSV = ({
-  data = null,
+export const exportToCSV = (
+  data: ExportOutput = [],
   columnDelimiter = ',',
-  lineDelimiter = '\n',
-}: {
-  data?: Record<string, any>;
-  columnDelimiter?: string;
-  lineDelimiter?: string;
-}) => {
-  let result: string;
-  let ctr: number;
-
-  if (data === null || !data.length) {
-    return null;
-  }
-
-  const keys = Object.keys(data[0]);
-
-  result = '';
-  result += keys.join(columnDelimiter);
-  result += lineDelimiter;
-
-  data.forEach((item: Record<string, any>) => {
-    ctr = 0;
-    keys.forEach((key) => {
-      if (ctr > 0) {
-        result += columnDelimiter;
-      }
-
-      let value = item[key];
-
-      if (
-        typeof item[key] === 'string' &&
-        item[key].includes(columnDelimiter)
-      ) {
-        value = `"${item[key]}"`;
-      }
-
-      if (Array.isArray(item[key])) {
-        value = `"${item[key].join(', ')}"`;
-      }
-
-      result += value;
-      ctr++;
-    });
-    result += lineDelimiter;
+  lineDelimiter = '\n'
+) => {
+  let result = '';
+  data.forEach((row: Array<string | number>) => {
+    result += row.join(columnDelimiter) + lineDelimiter;
   });
 
   return result;