+++ packages/widget-picker/CHANGELOG.md
@@ -3,6 +3,97 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.4](https://github.com/keen/keen/compare/@keen.io/widget-picker@1.4.3...@keen.io/widget-picker@1.4.4) (2021-09-30)
+
+**Note:** Version bump only for package @keen.io/widget-picker
+
+
+
+
+
+## [1.4.3](https://github.com/keen/keen/compare/@keen.io/widget-picker@1.4.2...@keen.io/widget-picker@1.4.3) (2021-09-27)
+
+**Note:** Version bump only for package @keen.io/widget-picker
+
+
+
+
+
+## [1.4.2](https://github.com/keen/keen/compare/@keen.io/widget-picker@1.4.1...@keen.io/widget-picker@1.4.2) (2021-09-24)
+
+**Note:** Version bump only for package @keen.io/widget-picker
+
+
+
+
+
+## [1.4.1](https://github.com/keen/keen/compare/@keen.io/widget-picker@1.4.0...@keen.io/widget-picker@1.4.1) (2021-09-24)
+
+**Note:** Version bump only for package @keen.io/widget-picker
+
+
+
+
+
+# [1.4.0](https://github.com/keen/keen/compare/@keen.io/widget-picker@1.3.16...@keen.io/widget-picker@1.4.0) (2021-09-17)
+
+
+### Features
+
+* 🎸 add support for multi-analysis ([2249ccf](https://github.com/keen/keen/commit/2249ccf7a28d9102741847d9f6f7aa671c26cbe7))
+
+
+
+
+
+## [1.3.16](https://github.com/keen/keen/compare/@keen.io/widget-picker@1.3.15...@keen.io/widget-picker@1.3.16) (2021-09-16)
+
+**Note:** Version bump only for package @keen.io/widget-picker
+
+
+
+
+
+## [1.3.15](https://github.com/keen/keen/compare/@keen.io/widget-picker@1.3.14...@keen.io/widget-picker@1.3.15) (2021-09-14)
+
+**Note:** Version bump only for package @keen.io/widget-picker
+
+
+
+
+
+## [1.3.14](https://github.com/keen/keen/compare/@keen.io/widget-picker@1.3.13...@keen.io/widget-picker@1.3.14) (2021-09-10)
+
+**Note:** Version bump only for package @keen.io/widget-picker
+
+
+
+
+
+## [1.3.13](https://github.com/keen/keen/compare/@keen.io/widget-picker@1.3.12...@keen.io/widget-picker@1.3.13) (2021-09-09)
+
+**Note:** Version bump only for package @keen.io/widget-picker
+
+
+
+
+
+## [1.3.12](https://github.com/keen/keen/compare/@keen.io/widget-picker@1.3.11...@keen.io/widget-picker@1.3.12) (2021-09-09)
+
+**Note:** Version bump only for package @keen.io/widget-picker
+
+
+
+
+
+## [1.3.11](https://github.com/keen/keen/compare/@keen.io/widget-picker@1.3.10...@keen.io/widget-picker@1.3.11) (2021-09-09)
+
+**Note:** Version bump only for package @keen.io/widget-picker
+
+
+
+
+
 ## [1.3.10](https://github.com/keen/keen/compare/@keen.io/widget-picker@1.3.9...@keen.io/widget-picker@1.3.10) (2021-09-08)
 
 **Note:** Version bump only for package @keen.io/widget-picker